pyproject.toml
[project]
name = "minecraft-auto-angler"
version = "0.1.0"
description = "A toy project to automate fishing in Minecraft."
requires-python = ">=3.10"
dependencies = [
    "opencv-python~=4.7.0.68",
    "numpy>=1.26.0,<2.0",
    "pyautogui~=0.9.53",
    "Pillow~=9.4.0",
    "pynput~=1.7.6",
    "mss",
]

[project.scripts]
autoangler = "autoangler.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "autoangler/**",
]

[tool.hatch.build.targets.wheel]
packages = ["autoangler"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
